refactor: decentralize types and remove unused type definitions
Moves TypeScript types from centralized src/types/index.ts to co-located files:
- src/services/types.ts: RecordingState, EventListener, UnsubscribeFunction
- src/views/types.ts: RecordingUICallbacks
- src/components/types.ts: RecordingUIProps

Removes 9 unused type definitions including PluginEvent, RecordingEvent,
SyncCallback, AsyncCallback, PluginConfig, PluginEnvironment, UIElementOptions,
ButtonOptions, ContainerOptions, and RecordingTransition.

Updates all import statements across 5 source files and 1 test file.
This improves maintainability by co-locating types with their usage,
making them easier to discover and preventing forgotten unused types.
