+++ src/services/GlobalRecordingState.ts
@@ -1,4 +1,4 @@
-import type { EventListener, RecordingState, UnsubscribeFunction } from "../types/index.js";
+import type { EventListener, RecordingState, UnsubscribeFunction } from "./types.js";
 
 /**
  * Global recording state coordinator that prevents concurrent recordings across the plugin.
