+++ src/views/RecordingView.ts
@@ -2,7 +2,7 @@ import { ItemView, type WorkspaceLeaf } from "obsidian";
 import { RecordingUI } from "../components/RecordingUI.js";
 import type { GlobalRecordingState } from "../services/GlobalRecordingState.js";
 import { LocalRecordingState } from "../services/LocalRecordingState.js";
-import type { RecordingUICallbacks } from "../types/index.js";
+import type { RecordingUICallbacks } from "./types.js";
 
 export const RECORDING_VIEW_TYPE = "operator-recording-view";
